Extract a `TransformMixin` from `BaseOperator`
