+++ merlin/dag/__init__.py
@@ -14,8 +14,10 @@
 # limitations under the License.
 #
 
-# flake8: noqa
-from merlin.dag.base_operator import BaseOperator, Supports
+from merlin.dag.base_operator import BaseOperator
 from merlin.dag.graph import Graph
 from merlin.dag.node import Node, iter_nodes, postorder_iter_nodes, preorder_iter_nodes
 from merlin.dag.selector import ColumnSelector
+
+# flake8: noqa
+from merlin.dag.transform_mixin import Supports, TransformMixin
